基础开发| 1.书籍单本上传更新 2.书籍按id下载

+++ src/main/java/com/favor/book/dao/BookRepository.java
@@ -17,5 +17,5 @@ import org.springframework.stereotype.Repository;
  * 2.dao层继承Repository<实体类型,主键id的类型>
  */
 @Repository
-public interface BookRepository extends JpaRepository<Book, Integer>, QuerydslPredicateExecutor<Book> {
+public interface BookRepository extends JpaRepository<Book, Long>, QuerydslPredicateExecutor<Book> {
 }
